Updated a few doubles to be ints

+++ web/client-backplane/src/main/java/io/deephaven/javascript/proto/dhinternal/flatbuffers/Long.java
@@ -8,12 +8,12 @@ import jsinterop.annotations.JsType;
 public class Long {
     public static Long ZERO;
 
-    public static native Long create(double low, double high);
+    public static native Long create(int low, int high);
 
-    public double high;
-    public double low;
+    public int high;
+    public int low;
 
-    public Long(double low, double high) {}
+    public Long(int low, int high) {}
 
     @JsMethod(name = "equals")
     public native boolean equals_(Object other);
